+++ Thunderstore/CHANGELOG.md
@@ -1,3 +1,9 @@
+## Version 1.3.0
+
+- Include utility slot in teleport logic
+- Added `[utilslot]` and `[nonutilslot]` filters
+- Fix issue displaying weight after teleporting
+
 ## Version 1.2.4
 
 - Compatibility update for v80